move extract.py to python module
