+++ backend/corpora/dbnl/dbnl.py
@@ -5,7 +5,7 @@ from tqdm import tqdm
 
 from django.conf import settings
 from addcorpus.python_corpora.corpus import XMLCorpusDefinition, FieldDefinition
-from addcorpus.extract import Metadata, XML, Pass, Order, Backup, Combined
+from addcorpus.python_corpora.extract import Metadata, XML, Pass, Order, Backup, Combined
 import corpora.dbnl.utils as utils
 from addcorpus.es_mappings import *
 from addcorpus.python_corpora.filters import RangeFilter, MultipleChoiceFilter, BooleanFilter
