+++ backend/corpora/dbnl/dbnl_metadata.py
@@ -1,7 +1,7 @@
 import os
 from django.conf import settings
 from addcorpus.python_corpora.corpus import CSVCorpusDefinition, FieldDefinition
-from addcorpus.extract import CSV, Combined, Pass
+from addcorpus.python_corpora.extract import CSV, Combined, Pass
 import corpora.dbnl.utils as utils
 
 class DBNLMetadata(CSVCorpusDefinition):
