+++ backend/corpora/dutchannualreports/dutchannualreports.py
@@ -7,7 +7,7 @@ from datetime import datetime
 
 from django.conf import settings
 
-from addcorpus.extract import XML, Metadata, Combined
+from addcorpus.python_corpora.extract import XML, Metadata, Combined
 from addcorpus.python_corpora.filters import MultipleChoiceFilter, RangeFilter
 from addcorpus.python_corpora.corpus import XMLCorpusDefinition, FieldDefinition
 from media.image_processing import get_pdf_info, retrieve_pdf, pdf_pages, build_partial_pdf
