+++ backend/corpora/dutchnewspapers/dutchnewspapers_public.py
@@ -12,7 +12,7 @@ from django.conf import settings
 
 from addcorpus.python_corpora.corpus import XMLCorpusDefinition, FieldDefinition, consolidate_start_end_years
 from addcorpus.python_corpora import filters
-from addcorpus.extract import Metadata, XML
+from addcorpus.python_corpora.extract import Metadata, XML
 from addcorpus.python_corpora.load_corpus import corpus_dir
 
 from corpora.utils.constants import document_context
