+++ backend/corpora/parliament/denmark-new.py
@@ -5,7 +5,7 @@ import os
 import re
 
 from corpora.parliament.parliament import Parliament
-from addcorpus.extract import Constant, CSV, Metadata, Combined
+from addcorpus.python_corpora.extract import Constant, CSV, Metadata, Combined
 from addcorpus.python_corpora.corpus import CSVCorpusDefinition
 import corpora.parliament.utils.field_defaults as field_defaults
 import corpora.utils.formatting as formatting
