+++ backend/corpora/parliament/denmark.py
@@ -4,7 +4,7 @@ import logging
 from django.conf import settings
 
 from corpora.parliament.parliament import Parliament
-from addcorpus.extract import Constant, CSV
+from addcorpus.python_corpora.extract import Constant, CSV
 from addcorpus.python_corpora.corpus import CSVCorpusDefinition
 import corpora.parliament.utils.field_defaults as field_defaults
 import corpora.utils.formatting as formatting
