+++ backend/corpora/parliament/finland.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from glob import glob
 
 from addcorpus.python_corpora.corpus import XMLCorpusDefinition
-from addcorpus.extract import XML, Combined, Constant, Metadata
+from addcorpus.python_corpora.extract import XML, Combined, Constant, Metadata
 from corpora.parliament.parliament import Parliament
 import corpora.parliament.utils.field_defaults as field_defaults
 from corpora.utils.constants import document_context
