+++ backend/corpora/parliament/germany-old.py
@@ -3,9 +3,9 @@ from datetime import datetime
 
 from django.conf import settings
 
-from parliament.parliament import Parliament
-from addcorpus.extract import Constant, CSV
-from backend.addcorpus.python_corpora.corpus import CSVCorpusDefinition
+from corpora.parliament.parliament import Parliament
+from addcorpus.python_corpora.extract import Constant, CSV
+from addcorpus.python_corpora.corpus import CSVCorpusDefinition
 import corpora.parliament.utils.field_defaults as field_defaults
