+++ backend/corpora/parliament/netherlands.py
@@ -7,7 +7,7 @@ from django.conf import settings
 
 import bs4
 from addcorpus.python_corpora.corpus import XMLCorpusDefinition
-from addcorpus.extract import XML, Constant, Combined, Choice
+from addcorpus.python_corpora.extract import XML, Constant, Combined, Choice
 from corpora.parliament.utils.parlamint import extract_all_party_data, extract_people_data, extract_role_data, party_attribute_extractor, person_attribute_extractor
 from corpora.utils.formatting import format_page_numbers
 from corpora.parliament.parliament import Parliament
