+++ backend/corpora/parliament/norway-new.py
@@ -2,7 +2,7 @@ from glob import glob
 from datetime import datetime
 from django.conf import settings
 
-from addcorpus.extract import Combined, Constant, CSV
+from addcorpus.python_corpora.extract import Combined, Constant, CSV
 from addcorpus.python_corpora.corpus import CSVCorpusDefinition
 from corpora.parliament.parliament import Parliament
 import corpora.parliament.utils.field_defaults as field_defaults
