+++ backend/corpora/parliament/sweden-old.py
@@ -2,7 +2,7 @@ from glob import glob
 from datetime import datetime
 
 from addcorpus.python_corpora.corpus import CSVCorpusDefinition
-from addcorpus.extract import CSV, Constant
+from addcorpus.python_corpora.extract import CSV, Constant
 from corpora.parliament.parliament import Parliament
 import corpora.parliament.utils.field_defaults as field_defaults
 import corpora.utils.constants as constants
