+++ backend/corpora/parliament/sweden.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from glob import glob
 
 from addcorpus.python_corpora.corpus import CSVCorpusDefinition
-from addcorpus.extract import CSV, Constant
+from addcorpus.python_corpora.extract import CSV, Constant
 from corpora.parliament.parliament import Parliament
 import corpora.utils.formatting as formatting
 import corpora.parliament.utils.field_defaults as field_defaults
