+++ backend/corpora/parliament/uk.py
@@ -4,7 +4,7 @@ from datetime import datetime
 
 from django.conf import settings
 
-from addcorpus.extract import Constant, CSV
+from addcorpus.python_corpora.extract import Constant, CSV
 from addcorpus.python_corpora.corpus import CSVCorpusDefinition
 from corpora.parliament.parliament import Parliament
 import corpora.parliament.utils.field_defaults as field_defaults
