+++ backend/corpora/peaceportal/FIJI/fiji.py
@@ -6,7 +6,7 @@ import logging
 from django.conf import settings
 
 from addcorpus.python_corpora.corpus import XMLCorpusDefinition
-from addcorpus.extract import XML, Constant, Combined
+from addcorpus.python_corpora.extract import XML, Constant, Combined
 from corpora.peaceportal.peaceportal import PeacePortal, categorize_material, join_commentaries, get_text_in_language
 from corpora.utils.exclude_fields import exclude_fields_without_extractor
