+++ backend/corpora/utils/test_corpora_utils.py
@@ -1,5 +1,5 @@
 from addcorpus.python_corpora.corpus import FieldDefinition
-from addcorpus.extract import Constant
+from addcorpus.python_corpora.extract import Constant
 
 from corpora.utils import exclude_fields
