+++ backend/download/tests/mock_corpora/multilingual_mock_corpus.py
@@ -3,7 +3,7 @@ import os
 
 from addcorpus.python_corpora.corpus import FieldDefinition, CSVCorpusDefinition
 from addcorpus.es_mappings import keyword_mapping, text_mapping
-from addcorpus.extract import CSV
+from addcorpus.python_corpora.extract import CSV
 
 # Fake corpus class for unit tests
