+++ backend/visualization/tests/mock_corpora/small_mock_corpus.py
@@ -2,7 +2,7 @@ from datetime import datetime
 import os
 
 from addcorpus.python_corpora.corpus import FieldDefinition, CSVCorpusDefinition
-from addcorpus.extract import CSV
+from addcorpus.python_corpora.extract import CSV
 from addcorpus.es_mappings import date_mapping, keyword_mapping, main_content_mapping, text_mapping
 from addcorpus.es_settings import es_settings
